config: replace replace_api_config
...with a privileged and an unprivileged version.

The privileged one works as previously.
The unprivileged one uses the api user for both owner and
group and 0660 as permissions.
The point is to have a convenient method to ensure the user
is correct regardless of whether we're running the
`-manager` binary as root or connect via the client (or run
the `-manager` binary as api user)

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>