replace Remote enum with a single struct
We currently do not anticipate product specific data in there and
instead expect that to live somewhere else.
For instance, specific metrics collection settings would live in the
metrics configuration.
The downside is that this means we may have to do more clean-up when
removing remotes later, but for now, this might be easier.

This requires bug fixes to the API macro's 'Updater' handling.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>