ui: main menu: fix truncated menu item text
when using a flex layout (which the selection view does), one must use
the 'pwt-flex-fit' class to be properly sized, not the 'pwt-fit' class.
Otherwise the sizing calculation of elements is rather weird because
of the css interaction of flex layout with non-flex layout, resulting in
the flex sized container not getting enough space.

While at it, replace some related hardcoded class names with the pwt::css
struct/enum helpers

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>