server: api: pve: add scan endpoint
This is intended to be the endpoint which we query when adding a remote.

As credentials users and tokens are accepted, and it can automatically
create a token on the PVE side, where the secret gets returned here for
further use. Default is `true` for that.

For now, we don't return any more information about the nodes, such as
ip addresses or fqdn, but it shouldn't be hard to add such things later
as additional information.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>