ui: remotes: wizard: rework and implement missing feature and use it
this completely reworks the wizard to work with the new scan endpoint,
allowing for pbs remotes, and finalizing the summary page.

We now reuse the 'Remote' type instead of the 'ServerInfo' struct for
storing information from the scan call, which is only done when using
the wizard for pve.

The wizard steps are as follows:

* initial connection

  user enters information on how to connect to the remote
  on pve, a 'connect' must be done that fetches metadata via the
  '/pve/scan' api call

* connections (pve only)

  since one can has multiple nodes/connections to a pve cluster, this
  list will be pre-filled with the nodenames+fingerprints from the scan
  call + the initial connection (since we cannot reliably tell which
  node it is, and we know this must work)

  this is not shown for pbs, as those are not clustered anyway, so
  having multiple connections here makes a lot less sense.

* info

  here the user can fill out the remote id to use (pre-filled with the
  hostname or cluster name, depending if it's a pve cluster or single
  host)

  and whether the user want us to auto-generate a token from the current
  authid (with the name as input field)

* summary

  sum up the entered information for a final check by the user

replace the 'add edit window' with the wizard, but only comment out
the edit code so we can switch back fast if we need to.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>