ui: provide a remote list context for the whole app
instead of polling the remote list in the main menu only, do it in the
top level app and provide it as a yew::Context. That way, any Component
can access that if it needs or wants to, regardless of how many levels
deep it's nested in the component tree.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>