ui: dashboard: make better use of the status api call
instead of having one panel with a very dense count, split it up into
one panel per type, so one can more easily digest that information.

This moves the loading of /resource/status into the dashboard itself
so we can use that info for several panels.

The `ResourceStatusPanel` is now obsolete with that, so remove it.

The Panels use a dynamic flex layout, so it adapts to the screen size
automatically.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>