ui: use rust-lld for linking wasm
this extends the `override_dh_auto_configure` rules hook and adds a
dependency for properly linking wasm projects when packaging. the
hook adapts the existing rustflags for the `wasm32-unknown-unknown`
target so that it uses the rust-lld linker instead of the
x86_64-linux-gnu-gcc one. rust-lld should be symlink to the lld
linker used by the current rust version's llvm integration.

otherwise, builds fail with an error messages such as (abbreviated for
clarity):

```
error: linking with `x86_64-linux-gnu-gcc` failed: exit status: 1
  |
[..]
  = note: x86_64-linux-gnu-gcc: error: unrecognized command-line option '--target=wasm32-unknown-unknown'
          x86_64-linux-gnu-gcc: error: unrecognized command-line option '--target=wasm32-unknown-unknown'
```

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>