server: metric collection: apply rrd journal on server startup
When implementing I confused syncing (which is just an fdatasync on the
journal file) with applying (read journal entries and apply them to
the cached database). This would lead to gap in the returned data
of up to 30 minutes (the default commit interval) in case of a
restart of the server.

Moved the call to the rrd_cache::init function, because I think its a
better fit there.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>