introduce abstractions for remotes config and remote client creation
Up until now, direct coupling to the pdm_config::*::config() functions
made it very hard to write any tests for PDM code. The config()
functions read configuration files from hardcoded filesystem paths,
paths that are not accessible to regular users.

To enable any meaningful testing, we need abstractions for this
low-level configuration persistence layer. In tests, we need to be able
to use a different (or reconfigured) config layer so that we can run
tests as a regular user, inject any test-relevant configuration and
check for any side effects from the tests.

Additionally, we also need to abstract the PVE and PBS API clients,
because we don't want unit and integration to have a need for an
actual running PVE/PBS instance (that should only be the case for
end-to-end tests). A previous commit added the PveClient trait to the
pve-api-types crate. This allows us to eliminate the direct
dependency on the concrete client implementation in most places in the
code.

This commit abstracts remote config access behind
a trait `RemoteConfig`, and remote client creation behind another trait
`ClientFactory`. The existing implementations have been moved
to structs which implement these traits, `DefaultClientFactory` and
`DefaultRemoteConfig`. During API server startup we create
instances of these structs and dependency inject them into the
appropriate modules.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
[WB: update proxmox-api-types submodule]
Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>