server: add "fake-remote" feature
This commit adds the option to run proxmox datacenter manager with
faked remotes. This is achieved by providing alternative
implementations for the RemoteConfig, ClientFactory and PveClient traits,
generating fake data which is determined by a json config file.

The main purpose of this is to evaluate the limits of metric collection,
resource API and the UI under extreme conditions.

The feature is gated behind the "fake-remote" feature flag which is
disabled by default. if the feature is enabled, the api daemons
check if the PDM_FAKED_REMOTE_CONFIG is set. If it is set, the
daemons try to load the config for the 'fake remotes' from the
file passed in the env var.

In the config, one can configure the number of remotes, number of
resources per remote, as well as a delay for any fake api call.

At the moment, only the resource apis and metrics apis are implemented,
any other will simply throw an error if used.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>