api types: add subscription types
this introduces some api types we want to use for caching and returning
subscription information. We introduce new types here vs reusing from
pbs/pve api-types since we don't need every information and also
those types are currently not the same (they're mostly compatible though)

It would probably be nicer to be able to reuse some types or move some
types to the proxmox-subscription crate, but that's currently not easy
as that has a hard dependency on openssl/proxmox-sys which we can't use
in wasm (where we want to use these types too).

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>