server: api: resources: add new `get_subscription` api call
This uses caching in the same manner as the resources list, and caches
a list of node subscription status for each remote. It returns only an
overall state for the remote by default that is calculated from the node
states.

By default the 'max-age' parameter is set to 1 day, so we don't often
refresh that info. Ideally it is refreshed by the backend e.g. twice a
day (and on demand in the ui maybe).

When setting the 'verbose' parameter to true, it also returns the list
of keys and state, this will be useful later when we maybe introduce
some kind of subscription management.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Dietmar Maurer <dietmar@proxmox.com>