ui: main: use new subscription info api and wait for that on login
rework the logic on how to show the subscription box a bit, so that
1. uses the new api call that returns subscription info from all remotes
2. blocks the login window until the subscription api call is done

Without doing 2, the subscription popup could be shown much later while
the user may already has interacted with the page, which is rather
jarring.

We only show the subscription popup when there is a remote with a node
that has no subscription. If we cannot fetch the subscription info from
a node, that will *not* trigger the popup.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>