types/server/cli: introduce separate type for pve resources
instead of reusing the pve-api-types ClusterResource. This way the UI
and client, does not have to rely on the pve-api-types, but can use a
type from the pdm-api-types.

This makes it also easier to handle for the ui, as we don't have
to convert to another type too.

refactor the conversion in api/resources so we can reuse them here.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>