ui: factor out pve status icons and improve search resource tree
by omitting the now unnecessary status columns, since it's now shown
together with the type icon (like we do in pve). Also change the text
rendering to something more like we have in pve, so we can replace the
'name' column with a 'node' one.

This makes info in the panel a bit more cleaner and easier to parse.
Also we'll reuse the refactored methods later in the pve specific
panels too.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>