ui: implement initial PVE tree panel
this replaces the current PVE panel with rrd graphs with a tree panel
similar to what we have on PVE itself, albeit not with all types (for
now).

It has an action column to start/stop and a deep link action.
There is also a 'filter' box in the toolbar.

When clicking on one entry, the right side panel gets updated
(empty for now except the node panel).

We also render the tags here, so for that we need some small css
classes.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>