ui: widget: add pve migrate map component
This is a combined network/storage map field, that loads the
config of a guest and extracts the used storages/networks and provides a
drop down for each one of them.

the resulting value is a  list of 'type:source=target', that must
be processed on the form side correctly.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>