ui: pve: rework pve panel layout
by moving the load of the tree into the top level component, and give
it via an Rc to the child elements (so we don't have to load it multiple
times).

Also show the panels with a gap between them and add a fixed panel for
the remote to show various cluster wide infos that may be interesting.

Changes a bit the behaviour for assembling the tree structure, and how
we initially select items there.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>