ui: make Desktop default theme and remove Material for now
The desktop theme is a relatively nice evolution of our current
widespread ExtJS Crisp, making it default will show that we're working
on some refreshed style while avoiding big (visual) churn.

While we also have an implementation of the ExtJS Crisp theme, it's so
close to the original that it's recognized as that but still to
different in key details to be conceived as complete reimplementation,
giving off an Uncanny Valley effect. It can be still good to have up
for selection to see that one can choose different themes and we might
support some customization for those that have UI ricing as hobby to
be finally happy too.

Remove the, well, rather colourful/eye-striking Material based theme
as it would need (IMO!) much more polishing in small details to
provide some good advantages over the other two themes, and I think
it's much easier to polish the Desktop theme further, as that is under
our design language control, whereas the Material one is not and often
has high expectations to be fully Material 3 – something that not even
Google manages to get right. As that theme and any opinion about it
should not steal the spotlight of the new project.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>