ui: buildsys: use hack for shared dependencies
This probably needs rethinking, but as long as we depend on the shared
crates as relative path we need to have something like this.

ps. I tried various other things, like unifying the root and UI cargo
workspace, which is ugly as well as Cargo cannot cope with a missing
workspace member, and there's then a lot of coupling between UI and
backend, something that we wanted to avoid in the first place through
this separation.

I also tried declaring the lib/ folder as separate own workspace, that
worked well for the backend but workspace inheriting broke when
building from the UI workspace, tried some things like copying the
whole folder or a symlink from ui/lib -> lib but neither worked, maybe
I'm just missing something, but for now lets go for the hack.

Note: I did not adapt the DSC target yet, but definitively something
to look into to avoid having to much tech debt now already.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>