ui: pve: reorganize panels on the left side
instead of having two different panels for remote summary + tree,
combine them into one, but sepearate them with a horizontal line.

this makes it necessary to reorganize the panel/column layout a bit, and
move the title out to the top level component (otherwiset the
gaps/paddings are not right).

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>