ui: add simple favicon spelling PDM
The SVG is a 128x128 vector with PDM using the Lato font and converted
to a path to avoid rendering issues when the font is not available.

Further I added a media query to change the path fill color when the
color scheme of the user agent (browser decoration) changes – note
that this is not necessarily the same as the color scheme of our UI,
which has no effect on the user agent's decoration.

SVG Icons are at time of writing supported by all major browsers but
Safari, just ignore that for now, but we probably want a fallback
16x16 bitmap icon for those in the future.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>