ui: fixate width of navigation to 275px
Until now the width changed dynamically depending on the navigation
entries, but that is a bit odd and also looks slightly cramped.

Some padding might have improved the latter, but dynamic sizes
depending on content is not _that_ ideal for the main navigation bar
IMO. So fixate the width to 275px, which allows to fit all entries
including the longest possible remote ID, which is limited to 32
characters from the safe ID character set.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>