ui: use vectorized and natively color-scheme aware Proxmox logo
Not only does the logo now actually looks crisp, especially if the
desktop or browser is scaling things up (i.e., zoomed in) like for
HiDPI, the size is almost the same size as the black bitmap one and we
lost the white one, which was 4 times bigger, completely.

Only set the height on the element, the browser then figures out the
correct proportions.

The displayed image itself became a bit bigger, that was done on
purpose when adapting the viewBox/padding of the vector for the UI.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>