ui: fix sizing and rendering of CPU and memory icon
Move the style closer to the one from (ExtJS) widget toolkit, this
fixes the baseline but also the size compared to Font Awesome icons,
as 14px font-size really does not translated to 14 px icon size there,
they are normally in a 16 x 16 px bounding box.

This fixes rendering, avoiding some jarring view, as the icons where
fine-tuned to be displayed at 16 px or more.

For the background-size use 16px 16px explicitly, as there was a
rendering bug in older Chromium based browser – seems to be gone now,
but I'm not sure in what version and it doesn't hurt to keep this.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>