ui: pve: cancels loads when changing guest/node
by dropping the old AsyncPool on change.
Due to the nature of the react-like behaviour, when we change from
one element to another of the same type, the component is reused, thus
the old loads will still finish. To prevent that, replace the old
AsyncPool with a fresh one, dropping the old ones which cancels the
loads.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>