ui: pve: tree: make action colum wider
in 'relaxed' spacing mode, the actions don't fit in 150px. Ideally we'd
want to make this depend on the spacer setting (which determines the
size of the buttons), but we use a grid-column-template here which does
not accept css  'calc()' values.

Until we have a good way to do something like this, simply use more
space for now that fits all sizes).

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>