server: implement basic resource cache schedule
this updates the resource cache in the background every ~60 minutes.
The default `max_age` in the ui is still much lower, so there we'll
likely always update (currently), but the `top-entities` call also
re-uses this info to check which resources actually exist. So to avoid
a race that when the user opens the ui the first time after the daemon
is restarted, we fill the cache for the `top-entities` to consume.

In the future we can tune the `max-age` setting and the
COLLECTION_INTERVAL to better catch user behavior so we mostly
read from the cache for the dashboard.

Also we can reuse that cache for remote status display when we want.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>