ui: gzip-compress WASM at build-time to reduce transfer size dramatically
This reduces the amount of data that needs to be transferred to load
the main WASM file by a factor of roughly 4.89, from 16.66 MB to
3.41 MB.

Avoid the need to adapt the wasm-bindgen generated code that imports
the WASM code and fetch the module data manually to pipe it through a
DecompressionStream [0] and then wrap it in a Response object again
which is what __wbg_init supports as parameter.

Set the Content-Type explicitly to 'application/wasm' to allow the use
of the faster WebAssembly.instantiateStreaming API.

How fetch is awaited might be still some smaller thing to optimize,
it's a while since I looked at how it works, but even if it loads
everything in memory as written it's just 3.4 MB, really not _that_
much and the whole decompressed 16+ MB WASM thing has to live in
memory once decompressed anyway.

This increases the Lighthouse Performance points for the initial load
from 55 to 68 in my Chromium 131 instance here.
And on a simulated 4G connection it reduces the time between load
and first "contentful paint" from 14.9s to 1.8s.

[0]: https://developer.mozilla.org/en-US/docs/Web/API/DecompressionStream

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>