ui: pve remote: fix calculating subscription level
We cannot assign the rendered subscription inside the loop, as then
the comparison in the next loop iteration for can never work, as we
try to compare the raw level with the rendered level. I.e., this could
only work with a single node.

Move the rendering part after the loop and track the raw level inside
the loop. This definitively would profit from some dedicated types,
like an enum for the state, a type for the level that implements the
rendering, and so on.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>