api: remote migrate: fix handling of host:port as remote
the PVE remote migrate API expects those to be set as standalone parameters,
passing in a "host:port" tuple as hostname fails the API call.

we already use http::uri::Authority to parse this when setting up the remote,
so do the same here and default to port 8006 if none is set explicitly.

Reported-by: Markus Frank <m.frank@proxmox.com>
Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>