server/ui: don't return token secret in api response
as a stop-gap, simply return the empty string for the secret.
Later we can fix that up with a proper type e.g. like we do in
proxmox-backup with `RemoteWithoutPassword`, but that touches too many
places here currently, so this seems fine for now.

On the UI side it's enough to simply don't mark it required and
add a placeholder. If its empty, it won't get submitted anyway.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>