ui: restructure wizard to have a better flow
instead of entering all connection info on the first page, just require
the hostname/port (+fingerprint). Then on the second page, require an id
and either user/password/realm/new tokenname or existing token/secret.

the third page is then the endpoint list, and the summary stayed the
same.

This requires a bit of restructuring of the place and way we collect the
information.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>