priv. api: do not clean-up socket unconditionally on start-up
This must be only done when (newly) binding the socket, not, e.g., on
reload, when the priv. daemon passes the FD along to the new,
re-execed, process and so the original socket file should be kept, to
allow the unpriv. daemon to proxy to the priv.

The alternative would be to keep the FD open in the unpriv. daemon and
transparently try to re-connect there. But independent of that,
cleaning up the socket file only when really needed makes always more
sense.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>