ui: also show running remote tasks in 'running tasks' list
by doing a second api call for the remote tasks, and mixing in the
results into the list.

To correctly show the remote tasks in the list, use the 'render'
property to render the remote tasks differently.

On the OpenTask viewstate, use a different baseurl for the task viewer
for remote tasks.

To determine if a UPID is a remote one, try to parse it as a
`RemoteUpid`.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>