drop WebUrl type for now
Instead use a http::Uri directly.
Uri does not implement the serde traits or `UpdaterType`, so we need
`#[serde(with)]` and `#[updater(type)]` attributes. But with the
`#[serde(with)]` we can later also turn it into a property string with
a default_key (which might need extra handling anyway, since urls can
technically contain `=` which might confuse the property string
parser).

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>