server: api task status: actually return the UPID
Since the UPID is already a dedicated parameter to the API endpoint
handler fn it is not present in the generic parameter Value map
anymore. But as we used that parameter map to extract it for the
return value we always returned an empty UPID.

Simply use the already extracted dedicated upid parameter instead.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
 [TL: reword commit message for clarification]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>