server: cache pve node reachability and names
Add a `remote_cache` module to access the cache which maps
(remote,host) to a struct holding the node name and a "reachable"
boolean. This can e used to connect to specific nodes or skip nodes
known to be currently unreachable (although the latter part is not
currently implemented).

The ClientFactory trait gains helpers to connect to specific nodes.

The unprivileged API gets a `tasks` submodule with a task which
fills the node cache and watches the remote config for updates (via
the config version cache).

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>
Reviewed-by: Lukas Wagner <l.wagner@proxmox.com>