ui: use on_activate for buttons
While it's technically the same currently, 'on_activate' is the correct
interface from our widget-toolkit. 'onclick' behaves the same due to
browser behavior for buttons, but when seeing 'onclick' one could be
lead to use that also for e.g. ActionIcons, where 'onclick' and
'on_activate' is not the same. So for consistency, use on_activate for
buttons everywhere.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250606072353.638426-2-d.csapak@proxmox.com