ui: dashboard: introduce DashboardConfig
to hold the configuration of the dashboard. We'll later use this to
hold the user configuration.

Having it as a property does not make sense here since we don't set it
from outside the dashboard, and since it's not a generic component,
it's not likely to be reused elsewhere.

Saving it as a PersistentState also makes it possible to have it persist
across sessions when we're able to change it.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250606072720.664054-4-d.csapak@proxmox.com