ui: dashboard: add status row and configuration window
This adds a row above the dashboard that includes:
* a 'refresh now' button
* a 'last refreshed' label
* a 'edit dashboard config' button

The last will open an EditWindow to configure the max-age parameter and
the refresh interval.

The status row is factored out, so the timer to update the label only
affects that small part of it instead of having to redraw the whole
dashboard.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250606072720.664054-6-d.csapak@proxmox.com