remote tasks: implement improved cache for remote tasks
This commit adds a new implementation for a cache for remote tasks, one
that should improve performance characteristics in for pretty much all
use cases.

In general storage works pretty similar to the task archive we already
have for (local) PDM tasks.

root@pdm-dev:/var/cache/proxmox-datacenter-manager/remote-tasks# ls -l
total 40
-rw-r--r-- 1 www-data www-data     0 Mar 13 13:18 active
-rw-r--r-- 1 www-data www-data  1676 Mar 11 14:51 archive.1741355462.zst
-rw-r--r-- 1 www-data www-data     0 Mar 11 14:51 archive.1741441862.zst
-rw-r--r-- 1 www-data www-data  2538 Mar 11 14:51 archive.1741528262.zst
-rw-r--r-- 1 www-data www-data  8428 Mar 11 15:07 archive.1741614662.zst
-rw-r--r-- 1 www-data www-data 11740 Mar 13 10:18 archive.1741701062
-rw-r--r-- 1 www-data www-data  3364 Mar 13 13:18 archive.1741788270
-rw-r--r-- 1 www-data www-data     0 Mar 13 13:18 journal
-rw-r--r-- 1 www-data www-data   287 Mar 13 13:18 state

Tasks are stored in the 'active' and multiple 'archive' files. Running
tasks are placed into the 'active' file, tasks that are finished are
persisted into one of the archive files. The archive files are suffixed
with a UNIX timestamp which serves as a lower-bound for start times for
tasks stored in this file. Encoding this lower-bound in the file name
instead of using a more traditional increasing file index (.1, .2, .3)
gives us the benefit that we can decide in which file a newly arrived
tasks belongs from a single readdir call, without even reading the file
itself.

Older archive files are additionally compressed using zstd.

The size of the entire archive can be controlled by doing file
'rotation', where the archive file with the oldest timestamp is deleted
and a new file with the current timestamp is added. If 'rotation'
happens at fixed intervals (e.g. once daily), this gives us a
configurable number of days of task history. There is no direct cap for
the total number of tasks, but this could be easily added by checking
the size of the most recent archive file and by rotating early if a
threshold is exceeded.

The format inside the files is also similar to the local task archive,
with one task corresponding to one line in the file. One key difference
is that here each line is a JSON object, that is to make it easier to
add additional data later, if needed. The JSON object contains the tasks
UPID, status, endtime and starttime (the starttime is technically also
encoded in the UPID, but having it as a separate value simplified a
couple of things) Each file is sorted by the task's start time, the
youngest task coming first.

One key difference between this task cache and the local task archive is
that we need to handle tasks coming in out-of-order, e.g. if remotes
were not reachable for a certain time. To maintain the ordering in the
file, we have to merge the newly arrived tasks into the existing task
file. This was implemented in a way that avoids reading the entire file
into memory at once, exploiting the fact that the contents of the
existing file are already sorted. This allows to do a zipper/merge-sort
like approach (see MergeTaskIterator for details). The existing file is
only read line-by-line and finally replaced atomically.

Since we always need to rewrite the entire archive file, even if there
is only a single new task added, there is also a journal/WAL file.
When adding new tasks, they will be first appended to the journal file
and then later applied to the actual archive files. The journal is
applied after a certain time or if the journal file itself grows too
large.

The cache also has a separate state file, containing additional
information, e.g. cut-off timestamps for the polling task. Some of the
data in the statefile is technically also contained in the archive
files, but reading the state file is much faster.

This commit also adds an elaborate suite of unit tests for this new
cache. While adding some additional work up front, they paid off
themselves during development quite quickly, since the overall approach
for the cache changed a couple of times. The test suite gave me
confidence that the design changes didn't screw anything up, catching a
couple of bugs in the process.

Finally, some concrete numbers about the performance improvments.
Benchmarking was done using the 'fake-remote' feature. There were 100
remotes, 10 PVE nodes per remote. The task cache contained
about 1.5 million tasks.
                                              before         after
list of active tasks (*):                     ~1.3s         ~300µs
list of 500 tasks, offset 0 (**):             ~1.3s         ~1.5ms
list of 500 tasks, offset 1 million (***):    ~1.3s         ~200ms
list of 500 tasks, offset 0,
    2000 tasks in journal (****):                           ~4.5ms
Size on disk:                                 ~500MB        ~40MB

(*):  Requested by the UI every 3s
(**): Requested by the UI when visiting Remotes > Tasks
(***): E.g. when scrolling towards the bottom of 'Remotes > Tasks'
(****): e.g. when the journal has not been applied for a while. Reading tasks
       from the journal is a bit less efficient than from the task archive, since
       we have to fully load it into memory so that we can sort the tasks and
       also  remove potential duplicates

In the old implementation, the archive file was *always* fully
deserialized and loaded into RAM, this is the reason why the time needed
is pretty idential for all scenarios. The new implementation reads the
archive files only line by line, and only 500 tasks were loaded into RAM
at the same time. The higher the offset, the more archive lines/files we
have to scan, which increases the time needed to access the data. The
tasks are sorted descending by starttime, as a result the requests get
slower the further you go back in history.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>