remote tasks: add background task for task polling, use new task cache
This commits changes the remote task module as follows:

- Add a new background task for regular polling of task data
Instead of triggering fetching of task data from the `get_tasks` function,
which is usually called by an API handler, we move the fetching to a
new background task. The task fetches the latest tasks from all remotes
and stores them in the task cache in regular intervals (10 minutes).
The `get_tasks` function itself only reads from the cache.
The main rationale for this change is that for large setups, fetching
tasks from all remotes can take a *long* time (e.g. hundreds of remotes,
each with a >100ms connection - adds up to minutes quickly).
If we do this from within `get_tasks`, the API handler calling the
function is also blocked for the entire time.
The `get_tasks` API is called every couple of seconds by the UI the get
a list of running remote tasks, so this *must* be quick.

- Tracked tasks are also polled in the same background task, but with
a short polling delay (10 seconds). If a tracked task finishes,
a out-of-order fetch of tasks for a remote is performed to update
the cache with all task data from the finished task.

- Only finished tasks are requested from the remotes. This avoids a
foreign (as in, not started by PDM) running task to appear stuck in
the running state until the next regular task cache refresh.
The tracked task polling could be extended to also poll running foreign
tasks, but this is easy addition for the future.

- Tasks are now stored in the new improved task cache implementation.
This should make retrieving tasks much quicker and avoids
unneeded disk IO.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>