remote tasks: log error in case of task panic, instead of cancelling all tasks
Should hopefully never happen, but for cases like these we still want to
to persist the tasks from other nodes instead of failing completely.

Also remove the `set_node_failure` function for NodeFetchSuccessMap. For
panic'd tasks we don't have a straightforward way to get the node name,
and we only ever check for success any way, not failure.

Suggested-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>