ui: pve wizard: connect: reset later pages when form changes
not on connect. With this, we don't reset the later pages anymore
when we change between tabs without changing anything.

The first connetion check will now only be done on the first navigation
away from the page or when the form changed afterwards.

Same for the 'Settings' panel. Only on the first change from 'Settings'
to 'Endpoints' or when the settings change will the endpoint list be
updated.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Reviewed-by: Lukas Wagner <l.wagner@proxmox.com>
Tested-by: Lukas Wagner <l.wagner@proxmox.com>