metric collection: persist state after metric collection
The metric collection has to maintain some state, at the moment this
is only the most recent timestamp from all received metric data points.
We use this as a cut-off time when requesting metric from the remote, as
everything *older* than that should already be stored in the database.

Up until now, the state was only stored in memory, which means that it
was lost when the daemon restarted.

The changes in this commit makes the metric collection system
load/save the state from a file in
/var/lib/proxmox-datacenter-manager/metric-collection-state.json

The following data points are stored for every remote:
- most-recent-datapoint:
  Timestamp of the most recent datapoint, used as a cut-off when
  fetching new metrics
- last-collection:
  `last-collection` field saves the timestamp of the last *successful*
  metric collection for a remote as local timestamp.
- error:
  String representation of any error that occured in the last collection
  attempt

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Reviewed-by: Maximiliano Sandoval <m.sandoval@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250826135119.336510-5-l.wagner@proxmox.com